refactor: refactor controller for include responseEntity in routes

+++ src/main/java/dev/ocaina/usercontroller/Users/UserController.java
@@ -1,8 +1,12 @@
 package dev.ocaina.usercontroller.Users;
 
 
+import org.apache.catalina.User;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.net.http.HttpResponse;
 import java.util.List;
 
 @RestController
@@ -12,27 +16,44 @@ public class UserController {
     public UserController(UserService userService) {this.userService = userService;}
 
     @GetMapping("/all")
-    public List<UserDTO> getAll(){
-        return userService.getAll();
+    public ResponseEntity<List<UserDTO>> getAll(){
+        List<UserDTO> list = userService.getAll();
+        return ResponseEntity.ok(list);
     }
 
     @GetMapping("/{id}")
-    public UserDTO getById(@PathVariable Long id){
-        return userService.getById(id);
+    public ResponseEntity<?> getById(@PathVariable Long id){
+        UserDTO user = userService.getById(id);
+        if(user != null){
+            return ResponseEntity.ok(user);
+        }
+        return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                .body("Usuario não encontrado");
     }
 
     @PostMapping("/create")
-    public UserDTO post(@RequestBody UserDTO model){
-        return userService.create(model);
+    public ResponseEntity<String> post(@RequestBody UserDTO model){
+        UserDTO user = userService.create(model);
+        return ResponseEntity.status(HttpStatus.CREATED).body("Usuario Criado:" + user.getName());
     }
 
     @PutMapping("/update/{id}")
-    public UserDTO update(@PathVariable Long id, @RequestBody UserDTO model){
-        return userService.update(id,model);
+    public ResponseEntity<?> update(@PathVariable Long id, @RequestBody UserDTO model){
+        UserDTO user = userService.getById(id);
+        if(user != null){
+            userService.update(id,model);
+            return ResponseEntity.ok(user);
+        }
+         return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Usuario não encontrado");
     }
 
     @DeleteMapping("/delete/{id}")
-    public void delete(@PathVariable Long id){
-        userService.delete(id);
+    public ResponseEntity<String> delete(@PathVariable Long id){
+
+        if(userService.getById(id) != null){
+            userService.delete(id);
+            return ResponseEntity.ok("Usuario Deletado");
+        }
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Usuario não encontrado");
     }
 }
